refactor: scenario and data
