+++ tests/internal/licence-holder/licence-holder.spec.js
@@ -1,4 +1,4 @@
-import scenarioData from '../../support/scenarios/licence.scenario.js'
+import scenarioData from '../../support/scenarios/unregistered-licence.scenario.js'
 import { test, expect } from '../../support/fixtures.js'
 import { formatLongDate } from '../../support/helpers/date.helpers.js'
