+++ tests/support/scenarios/company-contact.scenario.js
@@ -1,4 +1,4 @@
-import licenceScenario from './licence.scenario.js'
+import unregisteredLicenceScenario from './unregistered-licence.scenario.js'
 import companyContactData from '../data/company-contact.data.js'
 import notificationData from '../data/notification.data.js'
 import { mergeByKey } from '../helpers/scenario.helpers.js'
@@ -7,29 +7,29 @@ export const title = 'Company contact'
 export const description = 'A licence, licence holder, company, a contact and notification data'
 
 export default function () {
-  const licence = licenceScenario()
+  const unregisteredLicence = unregisteredLicenceScenario()
 
-  const companyContact = companyContactData(licence)
+  const companyContact = companyContactData(unregisteredLicence)
 
-  const editCompanyContact = companyContactData(licence, {
+  const editCompanyContact = companyContactData(unregisteredLicence, {
     department: 'Test Contact Edit Alerts',
     email: 'test.contact.edit@example.com'
   })
-  const removeCompanyContact = companyContactData(licence, {
+  const removeCompanyContact = companyContactData(unregisteredLicence, {
     department: 'Test Contact Remove',
     email: 'test.contact.remove@example.com'
   })
-  const restoreCompanyContact = companyContactData(licence, {
+  const restoreCompanyContact = companyContactData(unregisteredLicence, {
     department: 'Test Contact Restore',
     email: 'test.contact.restore@example.com'
   })
 
   return mergeByKey(
-    licence,
+    unregisteredLicence,
     companyContact,
     editCompanyContact,
     removeCompanyContact,
     restoreCompanyContact,
-    notificationData(licence.licences[0].licenceRef, restoreCompanyContact)
+    notificationData(unregisteredLicence.licences[0].licenceRef, restoreCompanyContact)
   )
 }
