+++ tests/support/scenarios/unregistered-licence.scenario.js
